Update package script and setup file
This commit updates the package.ps1 script to copy the Bonjour installer to the package directory from the environment variable. It also updates the setup.iss file to change the license file name from LICENSE to NOTICE. Additionally, it enables the installation of the Bonjour service for Windows in the [Run] section of the setup.iss file.